Added the errorhandler, helpers and move the common function to helpers folder
